services/inference/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
joblib
numpy
scikit-learn
shap
